Update 100-prime_factor.c
we start the loop counter i from 2 instead of 3, and we check the condition i * i <= n instead of i < 12057. This is because if n is not a prime number, one of its prime factors must be less than or equal to the square root of n.

Inside the loop, we simplify the condition to check if n is divisible by i. If it is, we divide n by i until it is no longer divisible by i.

Finally, we print the value of n after the loop.